+++ .devcontainer/hooks/postStart.sh
@@ -14,7 +14,7 @@ fi
 if [[ -f "$KEY_FILE" ]]; then
     echo "🔓  Unlocking licensed assets using $KEY_FILE..."
     if [[ -n "$(git status --porcelain)" ]]; then
-        git stash save -q "pre-unlock"
+        git stash push -q "pre-unlock"
         git-crypt unlock <(cat "$KEY_FILE" | base64 -d)
         git stash pop -q || true
     else