+++ .devcontainer/hooks/onCreate.sh
@@ -3,6 +3,6 @@ set -euo pipefail
 
 WORKSPACE_DIR=$(realpath /workspaces/*)
 
-npm ci
+npm ci --prefix _devenv/
 npm ci --prefix "$HOME/twinejs"
 poetry install -P "$WORKSPACE_DIR/_devenv/spindler"