+++ _devenv/src/constants.ts
@@ -1,8 +1,8 @@
 // If you're running the engine locally (only private beta testers can do this),
 // set this to true. Otherwise, set it to false.
-const localDev = false;
+const localDev = true;
 
 // This is where the Get Lost engine lives.
 export const gameUrl = localDev
-  ? "http://localhost:5174"
+  ? "http://localhost:5173"
   : "https://game.getlost.gg/";